myhomework20/requirements.txt
django~=3.2.0
pillow
django-debug-toolbar
django-imagekit
django-bootstrap-v5
